Refactor SystemStatus into modular components with hooks and types
